refactor: model 패키지 구분
